mm3438 * Removing project not used * Moving examples to same project
