--- example/pom.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-   <modelVersion>4.0.0</modelVersion>
-
-   <packaging>jar</packaging>
-
-   <groupId>your.group.id</groupId>
-   <artifactId>your.artifact.id</artifactId>
-   <name>your name</name>
-   <description>your description</description>
-   <version>0.0.0</version>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.mnubo</groupId>
-            <artifactId>java-sdk-client</artifactId>
-            <scope>compile</scope>
-            <version>1.3.0</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.3</version>
-                <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
