+++ src/main/java/com/mnubo/java/sdk/client/Constants.java
