+++ src/main/java/com/mnubo/java/sdk/client/example/MyAppTest.java
@@ -12,7 +12,7 @@
  * ---------------------------------------------------------------------------
  */
 
-package app;
+package com.mnubo.java.sdk.client.example;
 
 import java.io.BufferedReader;
 import java.io.File;
