+++ src/main/java/com/mnubo/java/sdk/client/mapper/SmartObjectSerializer.java
