+++ src/main/resources/object_full_with_owner.json
