+++ src/main/resources/owner_full.json
